Fix encoding of (non-ref) GLOBs
In order to encode a GLOB, we have to take a reference to it and encode that
reference.  In the process, it inserts an entry into $seen for this reference
that only exists ephemerally, during the block that handles non-refs.

The next time through, that reference address can be re-used, and it
mistakenly thinks it's seen a bunch of stuff, but only because of this
re-used ephemeral reference.